requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
groq==0.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-core==0.2.37
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
langsmith==0.1.108
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
orjson==3.10.7
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
setuptools==74.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.6
